pyproject.toml
[project]
name = "generative-ai-python-2025-05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-genai>=1.17.0",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.5",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
